Benchmark tuning so we can get more consistent numbers on HT vs non-HT
With our benchmarks now running on more hardware we see they really gave threadly a massive advantage on hyperthreaded hardware compared to non-hyper threaded hardware.
This changes the benchmarks so that the amount of CPU used on tasks is way less, and thus works the scheduler harder.
This means we have more tasks, that are running for shorter periods.  In the new results threadly still has a definite advantage over java.util, but not as much.  Most importantly the results between HT and non-HT hardware seem to coorelate.