Start of a prototype with xchart to parse graph data
Lots more thought in how we want to structure our graphs.  Because of that we restructured how we do our benchmark groups.  Which is the reason for merging this into master now.  I have updated our already updated records to this new group classification.

As far as actually generating graphs, sadly this library has one MAJOR issue.  Unfortanetly for line graphs we can't give x positions aribtrary labels.  The must be either numbers or dates.