More KeyDistributedExecutorUniqueKey changes
The benchmark still is not completing, now hanging.
This keeps the back pressure on submission, but adds a poll for the run state too.
This also now just outputs the run count at the time of stop rather than waiting for submitted completion.  This is more of an execute throughput rather than a submit throughput (which is more useful).